# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Star:
    properties:
      completed_calculations:
        type: integer
      constellation:
        type: string
      creation_date:
        type: string
      finish_date:
        type: string
      form_date:
        type: string
      id:
        type: integer
      mod_id:
        type: integer
      name:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
...
host: 127.0.0.1
info:
  contact:
    email: Barinovartem383@gmail.com
    name: API Support
    url: https://vk.com/aaaaaaeaaaa
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: Energy Calculation System
  version: "1.0"
paths:
...
  /scopes:
    get:
      description: Get all scopes with optional search filter
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get scopes list
      tags:
      - Scope
  /star/{id}:
    get:
      description: Get detailed star information with all scope calculations
      parameters:
      - description: Star ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get star with calculations
      tags:
      - Star
  /star/active:
    get:
      description: Get current user's active star ID and calculation count
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get active star info
      tags:
      - Star
  /star/delete/{id}:
    delete:
      description: Delete star and redirect to scopes page
      parameters:
      - description: Star ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "303":
          description: Redirect to /scopes
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete star
      tags:
      - Star
  /star/edit/{id}:
    put:
      description: Update star name and constellation
      parameters:
      - description: Star ID
        in: path
        name: id
        required: true
        type: string
      - description: Star data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Star'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Edit star
      tags:
      - Star
  /star/finish/{id}:
    put:
      description: Complete or deny star (moderator only)
      parameters:
      - description: Star ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Status: completed or denied'
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Finish star
      tags:
      - Star
  /star/form/{id}:
    put:
      consumes:
      - application/json
      description: Change star status to formed and update calculations
      parameters:
      - description: Star ID
        in: path
        name: id
        required: true
        type: string
      - description: Calculation values
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Form star
      tags:
      - Star
  /stars:
    get:
      description: |-
        Get stars with role-based access control.
        - Moderators (modstatus=2): see all stars except "active" and "deleted" statuses
        - Users (modstatus=1): see only their own stars except "active" and "deleted" statuses
        - Guests (modstatus=0): access denied (401)
      parameters:
      - description: Status filter (applies to non-active, non-deleted stars)
        in: query
        name: status
        type: string
      - description: Start date filter (2006-Jan-02)
        in: query
        name: start_date
        type: string
      - description: End date filter (2006-Jan-02)
        in: query
        name: end_date
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Star'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get stars list
      tags:
      - Star
...
schemes:
- http
swagger: "2.0"